+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -6,6 +6,10 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* **Experimental (pre-release builds only):** (only in pre-release versions):
+  Add support for using environment variables as context propagation carriers.
+  ([#7174](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7174))
+
 ## 1.15.3
 
 Released 2026-Apr-21
@@ -31,9 +35,6 @@ Released 2026-Apr-21
 * Fixed `BaggagePropagator` to strip baggage properties (e.g. `;metadata`)
   from values when parsing the `baggage` header.
   ([#7009](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7009))
-
-* Add support for using environment variables as context propagation carriers.
-  ([#7174](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7174))
 
 ## 1.15.2